+++ test/extglob-ending-with-state-char.js
@@ -1,5 +1,5 @@
 var test = require('tap').test
-var minimatch = require('../')
+var minimatch = require('../').default
 
 test('extglob ending with statechar', function (t) {
   t.notOk(minimatch('ax', 'a?(b*)'))